+++ test/warnings/make.jl
@@ -101,7 +101,9 @@ julia> WarningTests.run_warnings_test("at-eval")
 │    1 !in 2
 │    #└────┘ ── extra tokens after end of expression
 └ @ Documenter
-┌ Warning: Invalid type of object in @eval in src/at-eval.md:19-21
+┌ Warning: In src/at-eval.md:18-20: `@eval` block has an unsupported keyword argument: typo
+└ @ Documenter
+┌ Warning: Invalid type of object in @eval in src/at-eval.md:24-26
 │ ```@eval
 │ "expanded_"*"eval"
 │ ```
@@ -136,6 +138,10 @@ julia> WarningTests.run_warnings_test("at-example")
 │    # Error @ none:1:2
 │    1 !in 2
 │    #└────┘ ── extra tokens after end of expression
+└ @ Documenter
+┌ Warning: In src/at-example.md:18-20: `@example` block has an unsupported keyword argument: typo
+└ @ Documenter
+┌ Warning: In src/at-example.md:18-20: `@example` block keyword argument `ansicolor` must be `true` or `false`; ignoring :blue
 └ @ Documenter
 [ Info: CrossReferences: building cross-references.
 [ Info: CheckDocument: running document checks.
@@ -195,8 +201,10 @@ julia> WarningTests.run_warnings_test("at-setup")
 [ Info: SetupBuildDirectory: setting up build directory.
 [ Info: Doctest: running doctests.
 [ Info: ExpandTemplates: expanding markdown templates.
-┌ Warning: failed to run `@setup` block in src/at-setup.md:9-11
-│ ```@setup
+┌ Warning: In src/at-setup.md:13-15: `@setup` block has an unsupported keyword argument: typo
+└ @ Documenter
+┌ Warning: failed to run `@setup` block in src/at-setup.md:18-20
+│ ```@setup with-error
 │ 1 !in 2
 │ ```
 │   exception =
@@ -205,6 +213,8 @@ julia> WarningTests.run_warnings_test("at-setup")
 │    1 !in 2
 │    #└────┘ ── extra tokens after end of expression
 │    in expression starting at string:1
+└ @ Documenter
+┌ Warning: In src/at-setup.md:23-25: `@setup` block requires a name.
 └ @ Documenter
 [ Info: CrossReferences: building cross-references.
 [ Info: CheckDocument: running document checks.